moved accordion to independent folder in components
